HACK: Disable almost all emission of runtime metadata
At least some of these I want to turn back on, but this is sufficient
for running BitPaint with only SIL optimizations and no LLVM
optimizations (so, everything's been converted to static dispatch, but
some metadata references might still exist). It's going to be tricky
because somewhere in the chain of LLVM->XCOFF->PEF we're unable to
deal with relative references and aliases to the middle of symbols.