Relative-reference functions through their transition vector
A function pointer on Classic Mac OS (and AIX) doesn't point directly
to the function's code; instead, it points to a _transition vector,_
which contains both the function's address and the address of the
_table of contents_ associated with that function, which is used to
access globals. LLVM handles that automatically for normal function
pointers, but we need to do it specially for relative references,
which go through ptrtoint.