Demote constants to mutable globals if relatively-referenced
Specifically, if relatively-referenced from an already-mutable global.
This is part of the restrictions on XCOFF relative references only
being statically resolved if they're within the same section.

This isn't really the right way to do this; the right thing would be
to explicitly put the constants in the same section as the mutable
globals but leave them as constants. Unfortunately, the XCOFF backend
doesn't yet support explicit sections on globals, even if they're not
custom sections. At the point when that's implemented (by me or the
IBM folks), this pass can change to assign those custom sections
instead of demoting constants to variables.